+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/metrics/functions/TransformNull.kt
@@ -9,10 +9,10 @@ import ru.yandex.money.tools.grafana.dsl.metrics.Metric
  * @author iryabtsev (Igor Ryabtsev)
  * @since 15.11.2018
  */
-class TransformNull(private val metric: Metric, private val transformValue: Int) : Metric {
+class TransformNull(private val metric: Metric, private val transformValue: Double) : Metric {
     override fun asString() = """transformNull(${metric.asString()}, $transformValue)"""
 }
 
-infix fun Metric.transformNull(transformValue: Int) = TransformNull(this, transformValue)
+infix fun Metric.transformNull(transformValue: Double) = TransformNull(this, transformValue)
 
-infix fun String.transformNull(transformValue: Int) = TransformNull(StringMetric(this), transformValue)
+infix fun String.transformNull(transformValue: Double) = TransformNull(StringMetric(this), transformValue)
